+++ src/main/java/pe/edu/upc/center/backendNutriSmart/profiles/interfaces/rest/resources/UserResource.java
@@ -1,11 +1,11 @@
 package pe.edu.upc.center.backendNutriSmart.profiles.interfaces.rest.resources;
 
 public record UserResource(
-        Long id,
+        int id,
         String name,
         String email,
         String password,
         Boolean isActive,
         String birthDate,
-        Long userProfileId
+        int userProfileId
 ) {}
